Revert "Revert "Add webpack; serve HTML""
