+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "asteroids",
+  "version": "1.0.0",
+  "description": "This game was created Doug McInnes. His code can be found [here](https://github.com/dmcinnes/HTML5-Asteroids), and you can play his version of the game online at his website [here](http://dougmcinnes.com/2010/05/12/html-5-asteroids/).",
+  "main": "game.js",
+  "scripts": {
+    "dev": "webpack-dev-server --open --config webpack.config.js"
+  },
+  "repository": "git+https://github.com/salarchitetto/asteroids.git",
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/salarchitetto/asteroids/issues"
+  },
+  "homepage": "https://github.com/salarchitetto/asteroids#readme",
+  "devDependencies": {
+    "clean-webpack-plugin": "^0.1.18",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "html-webpack-plugin": "^2.30.1",
+    "uglifyjs-webpack-plugin": "^1.1.8",
+    "webpack": "3.11.0",
+    "webpack-dev-server": "2.11.1"
+  },
+  "dependencies": {
+    "jquery": "^1.11.0"
+  }
+}
